+++ dash_pydantic_form/fields/__init__.py
@@ -1,13 +1,11 @@
-import logging
 from datetime import date, time
 from enum import Enum
-from types import UnionType
-from typing import Literal, Union, get_args, get_origin
+from typing import Literal, get_origin
 
 from pydantic import BaseModel
 from pydantic.fields import FieldInfo
 
-from dash_pydantic_form.utils import get_non_null_annotation, is_subclass
+from dash_pydantic_form.utils import Type, get_non_null_annotation, is_subclass
 
 from . import all_fields as fields
 from .base_fields import BaseField, VisibilityFilter
@@ -30,22 +28,16 @@ DEFAULT_REPR = fields.Json
 def get_default_repr(field_info: FieldInfo, **kwargs) -> BaseField:
     """Get default field representation."""
     ann = get_non_null_annotation(field_info.annotation)
+    type_ = Type.classify(field_info.annotation, discriminator=field_info.discriminator)
 
-    # Test for model list
-    if get_origin(ann) == list and is_subclass(get_args(ann)[0], BaseModel):
+    if type_ == Type.SCALAR:
+        return DEFAULT_FIELDS_REPR.get(ann, DEFAULT_REPR)(**kwargs)
+
+    if type_ in [Type.MODEL, Type.DISCRIMINATED_MODEL]:
+        return fields.Model(**kwargs)
+
+    if type_ == Type.MODEL_LIST:
         return fields.ModelList(**kwargs)
-
-    # Test for discriminated model
-    if field_info.discriminator:
-        if get_origin(ann) in [Union, UnionType] and all(is_subclass(x, BaseModel) for x in get_args(ann)):
-            # return fields.DiscriminatedModel(**kwargs)
-            logging.info("Discriminated model fields not yet supported.")
-        else:
-            logging.info(f"Discriminator not supported for {field_info.annotation}")
-
-    # Test for simple types
-    if ann in DEFAULT_FIELDS_REPR:
-        return DEFAULT_FIELDS_REPR[ann](**kwargs)
 
     # Test for type origin
     origin = get_origin(ann)