+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Discriminated Model union #2
 
 ## [0.1.2] - 2024-05-16
 ### Added
@@ -12,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fields auto-generated docstrings
 
 ## Changed
-- Allow Python>=3.10,<3.13
+- Allow Python>=3.10,<3.13 #1
 
 ## [0.1.0] - 2024-05-13
 ### Added